El Niño and La Niña Data
This is the temperature of the water during the 1985 to 2020 years which determines whether there was a El Niño or La Niña was in effect. El Niño is when the water is still warmer than usual and La Niña is when the water is cooler than usual during certain times of the year.